Use interwiki links where possible for InstantCommons users
This results in substantially smaller pages, and feels a bit more
elegant.  It can't work in all cases, though, because InstantCommons
users don't necessarily appear in Commons' interwiki map.

Pywikibot has native support for generating interwiki links, but it
doesn't work in the general case.  We might be able to construct the
links automatically by reversing the interwiki map, but just taking an
interwiki prefix as part of the configuration for a user is a lot
simpler.  It's not like we need to track rapid updates to the interwiki
map anyway.