Remove namespace prefix from foreign file titles
The file titles returned by the API on other wikis are the localised
ones for those wikis, which means that they might not work on
Commons.  Since <gallery> doesn't require the namespace prefix, I
simply remove the namespace prefix from all the file titles imported
from foreign wikis.

I haven't done this at a higher level because that would involve
questions of what to do when a back-end returns files that differ
only in the namespace prefix (case and/or spacing).  InstantCommons
can't do this, but OpenStreetMap can.