Be more careful of "File:" prefixes when generating galleries
If an external reference is to "File:File:Example.jpg", it's not safe
to remove "File:" from the start when writing it to a gallery.
Usually that's safe, but in this case it leaves something that refers
to a different file.  I had thought this didn't matter because Commons
doesn't include any files named like that, but of course external
sites can refer to files that don't exist on Commons so we need to
handle them sensibly.