Have the various bot modes explicitly select a site
Since they specify a target page name, they're specific to their sites
anyway, so it's silly to require the caller to provide the correct one,
either through user-config.py or -site.  This also requires passing the
correct site through various function calls so as not to depend on the
default site.