Canonicalise filenames in accordance with Unicode 9.0.0
Testing suggests that MediaWiki on Commons uses the casing table from
Unicode 8.0.0 or 9.0.0 for capitalising the first letters of page
titles.  Thus when Usage Bot capitalises the first letter of a title,
it should not apply case mappings that were introduced in later
versions of Unicode.

I've dealt with this by adding a list of the characters that have
gained case mappings in later Unicode versions, which is a bit ugly
but is the best I can think of.

Thanks to Ivanbranco for reporting the problem.