Use paging when fetching data from taginfo
Taginfo has recently started requiring the use of the paging system in
its API when fetching keys that have more than 1000 values.  Both
image=* and wikimedia_commons=* are way over that limit, so we need to
use paging.  The maximum page size is currently 999, so that's the
page size I'm using.